o Fixed a bug in the test, causing an overflow error.

+++ trunk/jtds/src/test/net/sourceforge/jtds/test/CSUnitTest.java
@@ -25,7 +25,7 @@ import java.io.*;
 
 /**
  * @author builder
- * @version $Id: CSUnitTest.java,v 1.14 2009-08-04 10:33:44 ickzon Exp $
+ * @version $Id: CSUnitTest.java,v 1.15 2009-11-06 08:25:52 ickzon Exp $
  */
 public class CSUnitTest extends TestBase {
     public CSUnitTest(String name) {
@@ -573,7 +573,7 @@ public class CSUnitTest extends TestBase {
         output.println("myreal is " + rs.getDouble("myreal"));
         output.println("myint is " + rs.getInt("myint"));
         output.println("mysmallint is " + rs.getShort("mysmallint"));
-        output.println("mytinyint is " + rs.getByte("mytinyint"));
+        output.println("mytinyint is " + rs.getShort("mytinyint"));
         output.println("mymoney is " + rs.getObject("mymoney"));
         output.println("mysmallmoney is " + rs.getObject("mysmallmoney"));
         output.println("mybit is " + rs.getObject("mybit"));
